Improve webpack config (separate dev/pro files)
